#36 Rename: 공백 제거
